fix: address Playwright review comments
